+++ src/raycast.h
@@ -8,7 +8,7 @@
 //void thread_pool_add_work(void* tp, void (*func)(void*), void* arg_var);
 //void thread_pool_destroy(void* tp);
 void draw_first_person_level(
-    u32* output, edit_wall_id* edit_id_buffer, float* z_buffer,
+    u32* output, edit_wall_id* edit_id_buffer, u16* z_buffer,
     int start_x, int end_x, 
     int flash_frame, 
     level* this_level,
